requirements.txt
flask
google-generativeai
python-dotenv
gunicorn
uvicorn
